Moving VoteData to common project and using it in MAUI/Blazor
