+++ TeamVote.Server/Pages/TeamVoteDemo.razor
@@ -1,4 +1,5 @@
 ﻿@page "/teamvote"
+@using TeamVoteCommon;
 
 <PageTitle>TeamVote</PageTitle>
 
@@ -6,13 +7,29 @@
 
 <p role="status">TeamVote Demo</p>
 
-<button class="btn btn-primary" @onclick="IncrementCount">Click me</button>
+<ul>
+	@foreach ( var vote in _votes )
+	{
+	   <li>
+		  <label>@vote.UserId</label>
+		  <label>@vote.VoteValue</label>
+	   </li>
+	}
+</ul>
+
+<input placeholder="Something todo" @bind="_newTodo" />
+<button @onclick="AddTodo">Add todo</button>
 
 @code {
-	private int currentCount = 0;
+	private List<VoteData> _votes = new();
+	private string? _newTodo;
 
-	private void IncrementCount()
+	private void AddTodo()
 	{
-		currentCount++;
+		if ( !string.IsNullOrWhiteSpace( _newTodo ) )
+		{
+			_votes.Add( new VoteData { UserId = _newTodo, VoteValue = 5 } );
+			_newTodo = string.Empty;
+		}
 	}
 }
