+++ TeamVote.Server/TeamVote.Server.csproj
@@ -12,4 +12,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Azure.SignalR" Version="1.4.0" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\TeamVoteCommon\TeamVoteCommon.csproj" />
+  </ItemGroup>
 </Project>
