+++ TeamVote.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TeamVote", "TeamVote\TeamVo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TeamVote.Server", "TeamVote.Server\TeamVote.Server.csproj", "{2BF78627-1B97-4D16-8BCC-C52F592BC956}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeamVoteCommon", "TeamVoteCommon\TeamVoteCommon.csproj", "{52A82F47-EDBC-4614-A0BC-CAD07772748C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -23,6 +25,10 @@ Global
 		{2BF78627-1B97-4D16-8BCC-C52F592BC956}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2BF78627-1B97-4D16-8BCC-C52F592BC956}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2BF78627-1B97-4D16-8BCC-C52F592BC956}.Release|Any CPU.Build.0 = Release|Any CPU
+		{52A82F47-EDBC-4614-A0BC-CAD07772748C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{52A82F47-EDBC-4614-A0BC-CAD07772748C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{52A82F47-EDBC-4614-A0BC-CAD07772748C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{52A82F47-EDBC-4614-A0BC-CAD07772748C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
