+++ TeamVote/MainViewModel.cs
@@ -1,53 +1,9 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
-using System.Xml.Linq;
+using TeamVoteCommon;
 
 namespace TeamVote;
-
-public partial class VoteData : ObservableObject
-{
-   [ObservableProperty]
-   public string _userId;
-
-   [ObservableProperty]
-   public string _voteDisplay;
-
-   private int _voteValue;
-   public int VoteValue
-   {
-      set
-      {
-         _voteValue = value;
-         if ( _voteValue < 0 )
-         {
-            VoteDisplay = "<None>";
-         }
-         else
-         {
-            VoteDisplay = HideVotes ? "?" : value.ToString();
-         }
-      }
-      get { return _voteValue; }
-   }
-
-   private bool _hideVotes;
-   public bool HideVotes
-   {
-      get => _hideVotes;
-      set
-      {
-         _hideVotes = value;
-
-         VoteValue = VoteValue;
-      }
-   }
-
-   public VoteData Clone()
-   {
-      return (VoteData)this.MemberwiseClone();
-   }
-}
 
 public partial class MainViewModel : ObservableObject
 {
