+++ TeamVote/TeamVote.csproj
@@ -53,4 +53,7 @@
 		<PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="7.0.0-rc.1.22427.2" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="7.0.0-rc.1.22426.10" />
 	</ItemGroup>
+	<ItemGroup>
+	  <ProjectReference Include="..\TeamVoteCommon\TeamVoteCommon.csproj" />
+	</ItemGroup>
 </Project>
